+++ devtools/doc-recipe/build.sh
@@ -1,6 +1,12 @@
 #!/usr/bin/env bash -x
-# remove the source, since we depend on the built (conda) version
-rm pyemma -r
+
+set -x
+
+# we want to install some dependencies via pip
+export PIP_IGNORE_INSTALLED=false
+export PIP_NO_INDEX=false
+
+python grep_active_notebooks.py
 
 pyemma_version=`python -c "import pyemma as e; print(e.version)"`
 export BUILD_DIR=${PREFIX}/v${pyemma_version}
@@ -17,8 +23,7 @@ if [[ -d /group/ag_cmb/pyemma_performance/unpublished ]]; then
 fi
 
 # install requirements, which are not available in conda
-cd doc
-pip install -r requirements-build-doc.txt
+pip install -vvv -r requirements-build-doc.txt
 
 make clean
 make ipython-rst
@@ -30,4 +35,3 @@ rm -rf $BUILD_DIR/doctrees
 
 # remove the deps from $PREFIX so we have only the docs left.
 pip uninstall -y -r requirements-build-doc.txt
-