+++ pyemma/coordinates/clustering/interface.py
@@ -22,7 +22,6 @@ Created on 18.02.2015
 @author: marscher
 '''
 
-from __future__ import absolute_import
 
 import os