+++ pyemma/coordinates/tests/test_readers.py
@@ -1,6 +1,4 @@
-from __future__ import absolute_import
 
-import six
 import unittest
 import tempfile
 import shutil
@@ -52,7 +50,7 @@ class GenerateTestMatrix(type):
         return type.__new__(mcs, name, bases, attr)
 
 
-class TestReaders(six.with_metaclass(GenerateTestMatrix, unittest.TestCase)):
+class TestReaders(unittest.TestCase, metaclass=GenerateTestMatrix):
     """
     trajectory lengths:
         - 5000