+++ pyemma/coordinates/tests/test_save_traj.py
@@ -24,7 +24,6 @@ the retrival via save_traj
 @author: gph82, clonker
 """
 
-from __future__ import absolute_import
 
 import unittest
 import os