+++ pyemma/msm/api.py
@@ -20,7 +20,6 @@ r"""User API for the pyemma.msm package
 
 """
 
-from __future__ import absolute_import
 from .estimators import MaximumLikelihoodHMSM as _ML_HMSM
 from .estimators import BayesianMSM as _Bayes_MSM
 from .estimators import BayesianHMSM as _Bayes_HMSM
@@ -58,7 +57,6 @@ __all__ = ['markov_model',
 # =============================================================================
 
 
-# TODO: show_progress is not documented
 @shortcut('its')
 def timescales_msm(dtrajs, lags=None, nits=None, reversible=True, connected=True, weights='empirical',
                    errors=None, nsamples=50, n_jobs=None, show_progress=True, mincount_connectivity='1/n',
@@ -119,6 +117,9 @@ def timescales_msm(dtrajs, lags=None, nits=None, reversible=True, connected=True
 
     n_jobs : int, optional
         how many subprocesses to start to estimate the models for each lag time.
+
+    show_progress : bool, default=True
+        whether to show progress of estimation.
 
     mincount_connectivity : float or '1/n'
         minimum number of counts to consider a connection between two states.