+++ pyemma/msm/tests/test_hmsm.py
@@ -21,7 +21,6 @@ Test MLHMM.
 
 """
 
-from __future__ import absolute_import
 
 import unittest
 import numpy as np