+++ pyemma/util/tests/test_config.py
@@ -19,10 +19,7 @@ Created on 11.06.2015
 
 @author: marscher
 '''
-
-from __future__ import absolute_import, print_function
-
-from six.moves import configparser
+import configparser
 import os
 import sys
 import unittest
@@ -174,7 +171,7 @@ class TestConfig(unittest.TestCase):
     def test_mute_progress(self):
         """ switch mute on shall turn off progress bars"""
         from pyemma._base.progress import ProgressReporterMixin
-        import mock
+        from unittest import mock
         rp = ProgressReporterMixin()
 
         self.config_inst.mute = True